new LOS alg; walls don't show, some mobs don't show, everything is horribly slow
