+++ level/level.h
@@ -8,7 +8,6 @@
 #include "../mob/mob.h"
 #include "../chemistry/chemistry.h"
 #include "../simulation/simulation.h"
-#include "../los/los.h"
 
 typedef struct Level {
     unsigned char **tiles;
@@ -33,8 +32,5 @@ item* level_pop_item(level *lvl, int x, int y);
 
 bool is_position_valid(level *lvl, int x, int y);
 bool move_if_valid(level *lvl, mobile *mob, int x, int y);
-
-bool line_of_sight(level *lvl, int a_x, int a_y, int b_x, int b_y);
-bool can_see(level *lvl, mobile *actor, int target_x, int target_y);
 
 #endif
